Fixed more int format string warnings

+++ Frameworks/AIUtilities Framework/Source/AIDateFormatterAdditions.m
@@ -365,7 +365,7 @@ static AIDateFormatterCache *sharedFormatterCache = nil;
 		daysString		= (days == 1)		? ONE_DAY		: [NSString stringWithFormat:MULTIPLE_DAYS, (long)days];
 		hoursString		= (hours == 1)		? ONE_HOUR		: [NSString stringWithFormat:MULTIPLE_HOURS, (long)hours];
 		minutesString	= (minutes == 1)	? ONE_MINUTE	: [NSString stringWithFormat:MULTIPLE_MINUTES, (long)minutes];
-		secondsString	= (seconds == 1)	? ONE_SECOND	: [NSString stringWithFormat:MULTIPLE_SECONDS, (long)seconds];
+		secondsString	= (seconds == 1)	? ONE_SECOND	: [NSString stringWithFormat:MULTIPLE_SECONDS, (double)seconds];
 	}
 
 	//assemble the parts
